Extract most of #formatItem to HtmlFormatter
