+++ src/formatter/html_formatter.js
@@ -1,10 +1,34 @@
 import FormatterBase from './formatter_base';
+import Tag from '../chord_sheet/tag';
+
+const SPACE = '&nbsp;';
 
 export default class HtmlFormatter extends FormatterBase {
   constructor() {
     super();
     this.dirtyLine = false;
     this.lineEmpty = true;
+  }
+
+  formatItem(item) {
+    if (item instanceof Tag) {
+      return;
+    }
+
+    let chords = item.chords.trim();
+    let lyrics = item.lyrics.trim();
+
+    if (chords.length || lyrics.length) {
+      if (chords.length > lyrics.length) {
+        chords += SPACE;
+      } else if (lyrics.length > chords.length) {
+        lyrics += SPACE;
+      }
+
+      this.outputPair(chords, lyrics);
+    }
+
+    this.dirtyLine = true;
   }
 
   formatMetaData(song) {
